Add 1 year lookback period for BP002 indicator
First, all current files and objects with a 5y lookback period (as specified in the QOF business rules) get renamed to include "_5y_" in the name.

Next, a new dictionary gets created where the only difference is the lookback period, 5years get changed to 1year.

All actions in the project yaml get added and the final output file from the action `join_measuers` gets a new column to indicate whether the results come from the 1y or 5y lookback period.